Added support for the .no-sublime-package file to prevent a package from being re-zipped during the install process

+++ package_control/package_manager.py
@@ -693,33 +693,44 @@ class PackageManager():
             # since we don't want to accidentally overwrite user changes
             os.remove(tmp_package_path)
 
-            # For ST3 we create a .sublime-package file and place it in the Installed packages folder
             if int(sublime.version()) >= 3000:
-                try:
-                    package_file = zipfile.ZipFile(tmp_package_path, "w",
-                        compression=zipfile.ZIP_DEFLATED)
-                except (OSError, IOError) as e:
-                    show_error(u'An error occurred creating the package file %s in %s.\n\n%s' % (
-                        package_filename, tmp_dir, unicode_from_os(e)))
-                    return False
 
-                package_dir_regex = re.compile('^' + re.escape(package_dir))
-                for root, dirs, files in os.walk(package_dir):
-                    paths = dirs
-                    paths.extend(files)
-                    for path in paths:
-                        full_path = os.path.join(root, path)
-                        relative_path = re.sub(package_dir_regex, '', full_path)
-                        if os.path.isdir(full_path):
-                            continue
-                        package_file.write(full_path, relative_path)
+                # A package may have files that need to be extracted to work properly.
+                # For the package to signal this to Package Control, it must
+                # contain a file named ".no-sublime-package" in the root.
+                no_package_file = os.path.join(package_dir, '.no-sublime-package')
+                if os.path.exists(no_package_file):
+                    unpacked_package_dir = self.get_package_dir(package_name)
+                    shutil.move(package_dir, unpacked_package_dir)
 
-                package_file.close()
-                package_file = None
+                # By default for ST3 we create a .sublime-package file and place it
+                # in the Installed packages folder
+                else:
+                    try:
+                        package_file = zipfile.ZipFile(tmp_package_path, "w",
+                            compression=zipfile.ZIP_DEFLATED)
+                    except (OSError, IOError) as e:
+                        show_error(u'An error occurred creating the package file %s in %s.\n\n%s' % (
+                            package_filename, tmp_dir, unicode_from_os(e)))
+                        return False
 
-                if os.path.exists(package_path):
-                    os.remove(package_path)
-                shutil.move(tmp_package_path, package_path)
+                    package_dir_regex = re.compile('^' + re.escape(package_dir))
+                    for root, dirs, files in os.walk(package_dir):
+                        paths = dirs
+                        paths.extend(files)
+                        for path in paths:
+                            full_path = os.path.join(root, path)
+                            relative_path = re.sub(package_dir_regex, '', full_path)
+                            if os.path.isdir(full_path):
+                                continue
+                            package_file.write(full_path, relative_path)
+
+                    package_file.close()
+                    package_file = None
+
+                    if os.path.exists(package_path):
+                        os.remove(package_path)
+                    shutil.move(tmp_package_path, package_path)
 
             else:
                 # We have to remove the pristine package too or else Sublime Text 2
